request: include the core capability for all method calls
The RFCs 8620 and 8621 have been updated with an errata:

> The "urn:ietf:params:jmap:core" capability represents this document
> and MUST be included in the "using" list.

All method calls MUST include the core capability. Some server
implementations reject the requests performed with go-jmap when it is
missing.

To avoid updating all methods, forcefully add the capability when
performing a request in Client.Do() if it is missing. Move the constant
definition to the root module to avoid import cycles.

Link: https://www.rfc-editor.org/errata/eid6606
Reported-by: Benoit Tellier <btellier@linagora.com>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Signed-off-by: sami <sami@appscode.com>